+++ common/constants.h
@@ -16,7 +16,7 @@
 // Software version
 #define SW_VERSION_MAJOR 1
 #define SW_VERSION_MINOR 0
-#define SW_VERSION_PATCH 5
+#define SW_VERSION_PATCH 6
 #define SW_VERSION_BUILD 0
 
 // Number of max sensors connected to the board